cleaning after the restructuration
