+++ src/nectarchain/README.md
@@ -1,7 +1,7 @@
 Environment variable needed: 
 export NECTARCAMDATA="path to local NectarCam data, it can contain fits.fz run files, WaveformsContainer or ChargeContainer FITS files"
 
-Environment variables which can be defined: 
+Environment variables which can be defined but are optional: 
 export NECTARCHAIN_TEST="path to test for nectarchain"
 export NECTARCHAIN_LOG="path to log for nectarchain"
 export NECTARCHAIN_FIGURES="path to log figures for nectarchain"
