--- src/nectarchain/calibration/__init__.py
@@ -1,2 +1,0 @@
-"""nectarchain command line tools.
-"""
