+++ src/nectarchain/makers/calibration/gain/FlatFieldSPEMakers.py
@@ -33,7 +33,7 @@ from inspect import signature
 
 from .gainMakers import GainMaker
 
-from ...container import ChargeContainer
+from ....data.container import ChargeContainer
 
 from .parameters import Parameter, Parameters
