+++ src/nectarchain/makers/core.py
@@ -1,0 +1,19 @@
+import logging
+logging.basicConfig(format='%(asctime)s %(name)s %(levelname)s %(message)s')
+log = logging.getLogger(__name__)
+log.handlers = logging.getLogger('__main__').handlers
+
+from abc import ABC, abstractmethod
+
+__all__ = ["GeneralMaker"]
+
+class GeneralMaker(ABC):
+    """Mother class for all the makers, the role of makers is to do computation on the data. 
+    """
+    @abstractmethod
+    def make(self, *args, **kwargs):
+        """
+        Abstract method that needs to be implemented by subclasses.
+        This method is the main one, which computes and does the work. 
+        """
+        pass
